+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- Don't generate types for parameters with an `[FromServices]` annotation (#95)
+
 ## [0.13.1] - 2024-11-11
 
 ### Added